requirements.txt
discord.py==2.7.1
yt-dlp>=2026.3.0
spotipy==2.26.0
python-dotenv==1.1.0
PyNaCl==1.5.0
